requirements.txt
 langchain 
 langchain-core 
 langchain-groq 
 python-dotenv
 pandas
 chromadb
 langchain-community
